# granoflow-local-http-api.openapi.yaml
openapi: 3.1.0
info:
  title: Granoflow Local HTTP API
  version: 0.2.0
servers:
  - url: http://127.0.0.1:{port}
    variables:
      port:
        default: "56789"
security:
  - bearerAuth: []
...
paths:
...
  /v1/commands:
    post:
      tags: [Commands]
      summary: Execute command envelope
      operationId: postCommands
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommandEnvelope"
      responses:
        "200":
          description: Command response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandResult"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "423":
          $ref: "#/components/responses/Error"
...
  /v1/backup/exports:
    post:
      tags: [Commands]
      summary: Export an App backup through the shared LocalBackupController
      operationId: postBackupExport
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [outputPath]
              properties:
                outputPath:
                  type: string
      responses:
        "200":
          description: Backup export command response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandResult"
  /v1/backup/imports/preview:
    post:
      tags: [Commands]
      summary: Preview an App backup package
      operationId: postBackupImportPreview
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [inputPath]
              properties:
                inputPath:
                  type: string
      responses:
        "200":
          description: Backup preview command response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandResult"
  /v1/backup/imports:
    post:
      tags: [Commands]
      summary: Restore an App backup through the shared LocalBackupController
      operationId: postBackupImportRestore
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [inputPath, confirm]
              properties:
                inputPath:
                  type: string
                secretFile:
                  type: [string, "null"]
                confirm:
                  type: boolean
                allowMissingAttachments:
                  type: boolean
                allowLargeAttachmentConversion:
                  type: boolean
      responses:
        "200":
          description: Backup restore command response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandResult"
  /v1/ai-agent/tools:
    get:
      tags: [Commands]
      summary: List AI agent tools exposed by the App service layer
      operationId: listAiAgentTools
      responses:
        "200":
          description: AI agent tool list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandResult"
        "401":
          $ref: "#/components/responses/Error"
        "423":
          $ref: "#/components/responses/Error"
  /v1/ai-agent/tasks/{id}/export:
    parameters:
      - $ref: "#/components/parameters/ResourceId"
    get:
      tags: [Commands]
      summary: Export a single-task AI request package
      operationId: exportAiAgentTask
      responses:
        "200":
          description: AI task request package
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandResult"
        "401":
          $ref: "#/components/responses/Error"
        "423":
          $ref: "#/components/responses/Error"
  /v1/ai-agent/tasks/validate:
    post:
      tags: [Commands]
      summary: Validate a single-task AI reply
      operationId: validateAiAgentTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [path]
              properties:
                path:
                  type: string
                  description: Absolute local APKG file path.
      responses:
        "200":
          description: AI task validation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandResult"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "423":
          $ref: "#/components/responses/Error"
  /v1/ai-agent/tasks/import:
    post:
      tags: [Commands]
      summary: Preview or import a single-task AI reply
      operationId: importAiAgentTask
      parameters:
        - name: dryRun
          in: query
          required: false
          schema:
            type: boolean
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [path]
              properties:
                path:
                  type: string
                  description: Absolute local APKG file path.
                dryRunId:
                  type: string
                  description: Optional client dry-run correlation id.
                skipCardsWithMissingMedia:
                  type: boolean
                  default: false
                  description: Confirm the existing missing-media choice.
                stripRemoteMedia:
                  type: boolean
                  default: false
                  description: Remove remote media tokens before importing.
      responses:
        "200":
          description: AI task import or preview result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandResult"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "423":
          $ref: "#/components/responses/Error"
  /v1/tasks:
    get:
      tags: [Resources]
      summary: List tasks
      operationId: listTasks
      responses:
        "200":
          description: Resource command result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandResult"
    post:
      tags: [Resources]
      summary: Create task
      operationId: createTask
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        "201":
          description: Resource created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandResult"
  /v1/tasks/{id}:
    parameters:
      - $ref: "#/components/parameters/ResourceId"
    get:
      tags: [Resources]
      summary: Get task
      operationId: getTask
      responses:
        "200":
          description: Resource command result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandResult"
    patch:
      tags: [Resources]
      summary: Update task
      operationId: updateTask
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Resource command result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandResult"
    delete:
      tags: [Resources]
      summary: Delete task
      operationId: deleteTask
      responses:
        "200":
          description: Resource command result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandResult"
  /v1/tasks/{id}/complete:
    parameters:
      - $ref: "#/components/parameters/ResourceId"
    post:
      tags: [Resources]
      summary: Complete task
      description: Marks a task completed through the App business service layer.
      operationId: completeTask
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskCompleteRequest"
      responses:
        "200":
          description: Task completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandResult"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "423":
          $ref: "#/components/responses/Error"
...
  /v1/projects:
    get:
      tags: [Resources]
      summary: List projects
      operationId: listProjects
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandResult"
    post:
      tags: [Resources]
      summary: Create project
      operationId: createProject
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandResult"
  /v1/projects/{id}:
    parameters:
      - $ref: "#/components/parameters/ResourceId"
    get:
      tags: [Resources]
      summary: Get project
      operationId: getProject
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandResult"
    patch:
      tags: [Resources]
      summary: Update project
      operationId: updateProject
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandResult"
    delete:
      tags: [Resources]
      summary: Delete project
      operationId: deleteProject
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandResult"
  /v1/milestones:
    get:
      tags: [Resources]
      summary: List milestones
      operationId: listMilestones
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandResult"
    post:
      tags: [Resources]
      summary: Create milestone
      operationId: createMilestone
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandResult"
  /v1/milestones/{id}:
    parameters:
      - $ref: "#/components/parameters/ResourceId"
    get:
      tags: [Resources]
      summary: Get milestone
      operationId: getMilestone
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandResult"
    patch:
      tags: [Resources]
      summary: Update milestone
      operationId: updateMilestone
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandResult"
    delete:
      tags: [Resources]
      summary: Delete milestone
      operationId: deleteMilestone
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandResult"
...
  /v1/review-card-decks:
    get:
      tags: [Resources]
      summary: List review card decks
      operationId: listReviewCardDecks
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandResult"
    post:
      tags: [Resources]
      summary: Create review card deck
      operationId: createReviewCardDeck
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandResult"
...
  /v1/review-card-decks/import/package/preview:
    post:
      tags: [Resources]
      summary: Preview review card deck package import
      operationId: previewReviewCardDeckPackageImport
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandResult"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API Token
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: string
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
...
    CommandEnvelope:
      type: object
      additionalProperties: true
      description: |
        Forwarded to the App command handler as-is.
    CommandResult:
      type: object
      additionalProperties: true
      description: |
        App command result envelope.
    TaskCompleteRequest:
      type: object
      additionalProperties: false
      properties:
        startedAt:
          type: string
          format: date-time
          description: Optional start timestamp to apply before completing the task.
        endedAt:
          type: string
          format: date-time
          description: Optional completion timestamp to apply before completing the task.
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            detail:
              type: [string, "null"]